docker/flask/requirements.txt
flask
redis
confluent_kafka
paho-mqtt